+++ src/pylingdocs/data/format_templates/html/basic/hooks/post_gen_project.py
@@ -10,7 +10,7 @@ except ImportError:  # pragma: no cover
 WEB_DIR = files("pylingdocs") / "data/web"
 
 
-for filename in ["examples.css", "examples.js"]:
+for filename in ["examples.css", "examples.js", "glossing.js"]:
     source = Path(WEB_DIR / filename)
     target = Path(".")
     copy(source, target)